Extract endpoints and test data into separate files
