+++ Tests/JetworkingTests/MockData/MockResponseBody.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct MockBody: Codable {
+    let foo1: String
+    let foo2: String
+}
+
+struct VoidResult: Codable {}
